return a shallow copy of input strings if equal input and output codes
I encountered a bug in a tool that uses this library.
It loops through a list of output language codes, and reuses the array of input strings.
When input and output codes are the same, it performs some special processing on the returned array of input strings.
After this occurs, subsequent iterations of the loop pass invalid input strings to the translate function.
Strictly speaking, the code responsible for this error belongs to the tool and not the library.
However, if this library doesn't pipe input directly through to output (both of which are passed by reference)..
then this bug wouldn't occur; it would be more responsible for the library to return a shallow copy to prevent this class of bug.